requirements.txt
azure-storage-blob==12.26.0
Django==5.2.6
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
openai==1.106.1
psycopg2==2.9.10
python-dotenv==1.1.1